requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pandas==2.1.4
